fixed b+remove switch
Not entirely clear on why this should be the case, but if the address of
the transaction result is removed by including it in the `make-node`
call on line 367 it is later referenced and triggers an error. This must
be because the node is still in use, but the code has not been examined
carefully enough to ensure this is not indicative of a latent bug in the
b+tree removal algorithm.