`compact-database` function with caveats
Compaction of a database object is working with one major issue. Because
the `:cache` and the `:data` are separate there is a potentially gap
where a read could occur between the resetting of the `:data` and the
resetting of the `:cache`. This is problematic because, after
compaction, the addresses of the nodes will have changed.

At first glance, one option is to lump them together in the same atom,
another is to use refs instead. That said, some testing is required to
see what happens if a file read is attempted during file copying. If
that throws an error, then it will be necessary to employ some mechanism
to block reads during the final stage of compaction; if this is done,
then the separate atom updates should no longer be necessary.

It merits mentioning that during compaction an "archive" file is created
which contains the pre-compaction database. This functionality will
ultimately be altered or removed entirely since saving all the previous
data is antithetical to the disk-space-reclamation intent of
compaction.