Use canonical paths in database.
When a database is opened (or closed) the provided path string is turned
into its canonical representation. This avoids problems with opening the
same database with different supplied paths.

This change presented a bug in the archival/backup functionality where
it failed with absolute paths. The `archive-files` function has since
been modified to take a `dir` parameter which creates the archive
relative to the full database path.

It might be a good idea to tighten up the filepath handling,
e.g. instead of using (str path "/nodes") using something like:
`(str path +file-path-delim+ "nodes")`. Some further consideration is
warranted.