`backup-compressor` & `backup-fn` arguments for db
An `:backup-compressor` argument is supplied to an opened database
instead of being defined in a global var. If it is supplied, the
resulting filepath (or error) will be passed to a function provided as
the optional `:backup-fn` parameter. Err keys are supplied in lieu of
exceptions as the future is not passed back for dereferencing.

Doc strings explaining this behavior and the semantics of
opening/closing databases were also added to the core namespace.

The user benchmark now includes bzipping the archive and printing out
the archive result.