Changed backup handling to separate concerns
Archiving files is/should-be orthagonal to the handling of backups. Now,
the archiving functionality has been moved to a separate namespace
`codax.backup`. The `codax.store` namespace now only permits a
`backup-fn` argument which, when supplied, causes the compaction process
to generate uniquely timestamped files which are then passed to the
provided `backup-fn`. This prevents the `store` namespace from making
assumptions about the underlying system (e.g. shell access and shell
program availability) which are unrelated to its purpose.