Version 1.0.0 SNAPSHOT - Major Changes
This is a huge update to the system. The most important point is that it
eliminates the majority of the `codax.core` functions in favor of a
small set of functions which ensure database consistency and conform
well to general map-access idioms. This is more in keeping with the goal
of providing disk-backed maps with semantics that diverge minimally from
their in-memory brethren.

The unsafe functions have all been removed as testing demonstrated they
only provided a marginal performance boost. The tests and benchmarks
have been updated to reflect this fact. As a bonus, this decreases the
API considerably and eliminates the majority of duplicated and/or
ambiguous functionality.

Critically, a latent datastore consistency issue has been
corrected. Though, thus far, unseen in practice, adding elements at a
path beneath existing elements leaves the data in an inconsisent
state. Take the sequence;

`(assoc-at! db [:a] 1)`
`(assoc-at! db [:a :b] 2)`

This would result in an error when reassemling the collection as `:a` is
assigned to a primitive value but needs to be a map (containing
`:b`). Now, paths are prechecked to ensure this cannot happen. If an
attempt is made an error is thrown, just as would be the case when
attempting `(assoc-in {:a 1} [:a :b] 2)`. This incurs what appears to be
a small, but meaninful, performance penalty, which testing indicates is
in the vicinity of ~15%. At present this appears to be a worthwhile
tradeoff and is in keeping with the essence of the library. Though, a
better strategy made need to be devised moving forward.

NOTE: This firmly breaks compatibility with the levelpath library. Code
attempting to make the change will likely have to be
refactored. Ideally, this will result in ongoing maintainability.