Accept java.time.Instant as a valid type for keys
Because it is possible to convert losslessly between Instant, ZonedDateTime and
OffsetDateTime, we could choose to put the latter two in the same hex tag, so
they may be ordered relative to each other. We would need to come up with a
string representation that includes the offset or the zone, and the right
encoder/decoder fns.

In that case, it should be possible to extend the string representation further
so joda time objects can also be compared, and so on for all DateTime objects
which map directly to epoch time.

The current implementation only allows java.time.Instant.