add dynamic compaction scheduling
Previously the database waited for 10,000 updates to an open database
connection before running compaction. If the database was closed
before 10,000 updates the counter would be reset. This led to some
databases which were opened and closed intermittently never getting
compacted. The high threshold also led to small datasets growing
disproportionately large, despite compactions for such datasets being
extremely cheap.

The compactor now operates on the relative size of the
in-memory (active) and on-disk node manifests. The threshold factor
is computed such that as the database grows larger, the ratio of
post/pre compaction filesize grows smaller. For example, a database
that has ~100kB of data might grow to ~800kB before compaction, but a
database that has ~20MB of data should only grow to ~60MB before
compacting. This factor may be tuned more finely as real world use
cases demand.

Using the manifests as a proxy measure means that compactions are
unaffected by the closing and reopening of databases. It has the added
advantage that no changes to the file formats are requirede so it is
completely backward-compatible.

This addresses issue #23.