simplify and update benchmarking
The `codax.bench.performance` benchmark was updated to remain
compatible with the store and print the extra information offered to a
`*monitor-metrics*` debugging function. It was also simplified but
should still reflect a reasonably realistic workload. For simplicity
and determinism in testing, the wordlist in `codax.bench.wordlist` has
been pre-shuffled.

`codax.bench.users` was outdated and removed.

The `tuning.txt` file was of limited relevance and has been removed.