upgradable transaction improvements, tests, and docs
This is a large commit, but the major thing is that it ensures that
stale upgrade transactions are detected and eliminated. This is not a
problem most of the time, but if it happens that a compaction has
occurred during the upgrade (between releasing the read-lock and
acquiring the write-lock) the transaction will no longer work and
scary inscrutable errors will be thrown. There may be other subtle
issues elsewhere so in all cases it is forbidden and will throw a
"Transaction Invalidated By Upgrade" error if attempted.

Other things

- made small but breaking changes to `with-updatable-transaction` api
  - detecting upgrade exceptions from `ex-data` map changed
    - no longer rely on `:cause`
    - check for the presence of `:codax/upgraded-transaction`
  - option `:throw-on-restart` has been changed to `:throw-on-upgrade`
  - when manually resuming a transaction after upgrade you must use
    - the tx object provided in the ex-data
- implemented the `try-upgrade` macro
- upgradable transaction documentation
  - updated
  - expanded
  - moved to separate file docs/upgradable-transactions.md
  - improved example code
- added more tests
  - example tests
  - fuzz tests
  - path conflict detection tests